Extract Rust gatekeeper library
